fix(mobile): change settings privacy + screen sharing
