--- mobile/src/features/Settings/ui/screens/ChangeApplicationSettingsScreen/ScreenShare/ScreenShare.ts
@@ -1,35 +1,0 @@
-import {useSyncStorageToState} from '@yoroi/common'
-
-import * as React from 'react'
-import {NativeModules} from 'react-native'
-
-import {isAndroid} from '~/kernel/constants'
-import {screenShareStorageKeyManager} from '~/kernel/storage/storages'
-
-const {FlagSecure} = NativeModules
-
-export const useInitScreenShare = () => {
-  const [screenShareEnabled] = useSyncStorageToState(
-    screenShareStorageKeyManager,
-  )
-  const [initialised, setInitialised] = React.useState(false)
-
-  React.useEffect(() => {
-    if (!isAndroid || initialised) return
-
-    changeScreenShareNativeSettingOnAndroid(screenShareEnabled)
-    setInitialised(true)
-  }, [screenShareEnabled, initialised])
-
-  return {initialised}
-}
-
-export const changeScreenShareNativeSettingOnAndroid = (
-  screenShareEnabled: boolean,
-) => {
-  if (screenShareEnabled) {
-    FlagSecure.deactivate()
-  } else {
-    FlagSecure.activate()
-  }
-}
