+++ mobile/src/features/Transactions/useCases/TxDetails/AssetList.tsx
@@ -4,7 +4,7 @@ import {atoms as a, useTheme} from '@yoroi/theme'
 import * as React from 'react'
 import {FlatList, Text, TouchableOpacity, View} from 'react-native'
 
-import {usePrivacyMode} from '~/features/Settings/ui/screens/ChangeApplicationSettingsScreen/PrivacyMode/usePrivacyMode'
+import {usePrivacyMode} from '~/features/Settings/hooks/usePrivacyMode'
 import {useSelectedWallet} from '~/features/WalletManager/hooks/useSelectedWallet'
 import {normalisePtId} from '~/kernel/helpers/normalisePtId'
 import {useStrings} from '~/kernel/i18n/useStrings'
